requirements.txt
mediapipe==0.10.14
numba==0.60.0
numpy==2.0.1
opencv-contrib-python==4.10.0.84
pillow==10.4.0
